refactor(core): move EditingContextFactory interface to commonMain
Splits the factory interface so commonTest can inject a mock/fake
implementation. JVM-only reflective createNew(Class, vararg) method stays
in jvmMain via JvmEditingContextFactory subtype that extends both
EditingContextFactory (commonMain) and ContextFactory. Koin registers both
types so JVM-agnostic code can inject EditingContextFactory while GUI and
simulation code inject JvmEditingContextFactory.

Unblocks Bucket 0 test infrastructure portability — ~30 jvmTest files
that currently depend on EditingContextFactory via Koin DI can now be
migrated to commonTest with the same Koin module.

No behavior change; no callers break.
