+++ core/src/jvmTest/kotlin/cz/vutbr/fit/interlockSim/context/navigation/PathReservationRegistryTest.kt
@@ -19,7 +19,7 @@ import assertk.assertions.isInstanceOf
 import assertk.assertions.isTrue
 import cz.vutbr.fit.interlockSim.context.DefaultSimulationContext
 import cz.vutbr.fit.interlockSim.context.EditingContext
-import cz.vutbr.fit.interlockSim.context.EditingContextFactory
+import cz.vutbr.fit.interlockSim.context.JvmEditingContextFactory
 import cz.vutbr.fit.interlockSim.context.SimulationContextFactory
 import cz.vutbr.fit.interlockSim.objects.tracks.DynamicTrackBlock
 import cz.vutbr.fit.interlockSim.testutil.KoinTestBase
@@ -46,7 +46,7 @@ import java.io.InputStream
  * @since Issue #292 Phase 2 Enhancement (Plan Phase 2)
  */
 class PathReservationRegistryTest : KoinTestBase() {
-	private val editingContextFactory: EditingContextFactory by inject()
+	private val editingContextFactory: JvmEditingContextFactory by inject()
 	private val simulationContextFactory: SimulationContextFactory by inject()
 
 	private lateinit var registry: PathReservationRegistry
