+++ core/src/jvmTest/kotlin/cz/vutbr/fit/interlockSim/context/navigation/PathReservationServiceTest.kt
@@ -26,7 +26,7 @@ import assertk.assertions.isTrue
 import cz.vutbr.fit.interlockSim.testutil.TestFixtures
 import cz.vutbr.fit.interlockSim.context.DefaultSimulationContext
 import cz.vutbr.fit.interlockSim.context.EditingContext
-import cz.vutbr.fit.interlockSim.context.EditingContextFactory
+import cz.vutbr.fit.interlockSim.context.JvmEditingContextFactory
 import cz.vutbr.fit.interlockSim.context.SimulationContextFactory
 import cz.vutbr.fit.interlockSim.context.SimulationEnvironment
 import cz.vutbr.fit.interlockSim.objects.cells.DynamicInOut
@@ -70,7 +70,7 @@ import java.io.InputStream
  * @since Issue #294 (Phase 2 of Issue #292)
  */
 class PathReservationServiceTest : KoinTestBase() {
-	private val editingContextFactory: EditingContextFactory by inject()
+	private val editingContextFactory: JvmEditingContextFactory by inject()
 	private val simulationContextFactory: SimulationContextFactory by inject()
 
 	private lateinit var simulationContext: DefaultSimulationContext
