+++ core/src/jvmTest/kotlin/cz/vutbr/fit/interlockSim/objects/paths/AbstractPathSetUpSemaphoresTest.kt
@@ -14,7 +14,7 @@ import assertk.assertions.isGreaterThanOrEqualTo
 import assertk.assertions.isTrue
 import cz.vutbr.fit.interlockSim.context.DefaultSimulationContext
 import cz.vutbr.fit.interlockSim.context.EditingContext
-import cz.vutbr.fit.interlockSim.context.EditingContextFactory
+import cz.vutbr.fit.interlockSim.context.JvmEditingContextFactory
 import cz.vutbr.fit.interlockSim.context.SimulationContextFactory
 import cz.vutbr.fit.interlockSim.context.navigation.TopologyNavigator
 import cz.vutbr.fit.interlockSim.objects.cells.DynamicRailSwitch
@@ -70,7 +70,7 @@ import java.util.concurrent.TimeUnit.SECONDS
 @Tag("integration-test")
 @Timeout(30, unit = SECONDS)
 class AbstractPathSetUpSemaphoresTest : KoinTestBase() {
-	private val editingContextFactory: EditingContextFactory by inject()
+	private val editingContextFactory: JvmEditingContextFactory by inject()
 	private val simulationContextFactory: SimulationContextFactory by inject()
 
 	/**
