+++ desktop-ui/src/test/kotlin/cz/vutbr/fit/interlockSim/integration/ContextLifecycleIntegrationTest.kt
@@ -22,7 +22,7 @@ import cz.vutbr.fit.interlockSim.context.ContextTransformer
 import cz.vutbr.fit.interlockSim.context.DefaultEditingContext
 import cz.vutbr.fit.interlockSim.context.DefaultSimulationContext
 import cz.vutbr.fit.interlockSim.context.EditingContext
-import cz.vutbr.fit.interlockSim.context.EditingContextFactory
+import cz.vutbr.fit.interlockSim.context.JvmEditingContextFactory
 import cz.vutbr.fit.interlockSim.context.SimulationContext
 import cz.vutbr.fit.interlockSim.context.SimulationProcessFactory
 import cz.vutbr.fit.interlockSim.objects.cells.InOut
@@ -58,7 +58,7 @@ import cz.vutbr.fit.interlockSim.objects.core.ContextPropertyChangeListener
 @DisplayName("Context Lifecycle Integration")
 @Tag("integration-test")
 class ContextLifecycleIntegrationTest : KoinTestBase() {
-	private val editingContextFactory: EditingContextFactory by inject()
+	private val editingContextFactory: JvmEditingContextFactory by inject()
 	private val transformer: ContextTransformer by inject()
 	private val processFactory: SimulationProcessFactory by inject()
