+++ desktop-ui/src/test/kotlin/cz/vutbr/fit/interlockSim/xml/XMLContextFactoryOutputStreamTest.kt
@@ -19,7 +19,7 @@ import assertk.assertions.isNotNull
 import assertk.assertions.isTrue
 import cz.vutbr.fit.interlockSim.util.Resources
 import cz.vutbr.fit.interlockSim.context.EditingContext
-import cz.vutbr.fit.interlockSim.context.EditingContextFactory
+import cz.vutbr.fit.interlockSim.context.JvmEditingContextFactory
 import cz.vutbr.fit.interlockSim.objects.cells.InOut
 import cz.vutbr.fit.interlockSim.objects.cells.RailSwitch
 import cz.vutbr.fit.interlockSim.objects.core.Cell
@@ -55,7 +55,7 @@ import java.util.concurrent.TimeUnit
  */
 @Timeout(value = 10, unit = TimeUnit.SECONDS)
 class XMLContextFactoryOutputStreamTest : KoinTestBase() {
-	private val editingContextFactory: EditingContextFactory by inject()
+	private val editingContextFactory: JvmEditingContextFactory by inject()
 
 	@Nested
 	@DisplayName("Basic OutputStream Serialization")
